Mirror examples

+++ docker/rm.images.sh
@@ -1,12 +1,22 @@
 #!/bin/bash
-echo Removing images
-echo -e "--- Your Docker Images ---"
+echo -e "------ Your Docker Images ------"
 docker images
-echo -en " Are you sure you want to remove all of them? [n]|y > "
+echo -e "--------------------------------"
+echo -en "Are you sure you want to remove all of them? [n]|y|p (p: prompted) > "
 read a
-if [ "$a" == 'Y' ] || [ "$a" == 'y' ]
+ONE_BY_ONE=false
+if [ "$a" == 'Y' ] || [ "$a" == 'y' ] || [ "$a" == 'P' ] || [ "$a" == 'p' ]
 then
-  echo -e "Ok, removing..."
+  if [ "$a" == 'P' ] || [ "$a" == 'p' ]
+  then
+    ONE_BY_ONE=true
+  fi
+  echo -en "Ok, removing"
+  if [ "$ONE_BY_ONE" == "true" ]
+  then
+    echo -en " one by one"
+  fi 
+  echo -e "..."
 else
   echo -e "Aborting"
   exit 1
@@ -14,6 +24,21 @@ fi
 docker images | grep -v IMAGE | awk '{ print $3 }' > images
 for iid in `cat images`
 do
-  docker rmi $iid
+  if [ "$ONE_BY_ONE" == "true" ]
+  then
+    IMG_NAME=`docker images | grep $iid | awk '{ print $1 }' `
+    echo -en ">> Remove image $IMG_NAME ($iid) [n]|y > "
+    read resp
+	if [ "$resp" == 'Y' ] || [ "$resp" == 'y' ]
+	then  
+	  echo -e "... Removing $iid"
+	  docker rmi $iid  
+	else
+	  echo -e "Leaving $IMG_NAME alone"	
+	fi
+  else
+    docker rmi $iid
+  fi
 done
+rm images > /dev/null
 echo Done.
